day tree --partial

+++ server/src/routes/memories.ts
@@ -3,8 +3,15 @@ import { z } from "zod";
 import { prisma } from "../lib/prisma";
 
 export async function memoriesRoutes(app: FastifyInstance) {
-  app.get("/memories", async () => {
+  app.addHook("preHandler", async (request) => {
+    await request.jwtVerify();
+  });
+
+  app.get("/memories", async (request) => {
     const memories = await prisma.memory.findMany({
+      where: {
+        userId: request.user.sub,
+      },
       orderBy: {
         createdAt: "asc",
       },
@@ -19,7 +26,7 @@ export async function memoriesRoutes(app: FastifyInstance) {
     });
   });
 
-  app.get("/memories/:id", async (request) => {
+  app.get("/memories/:id", async (request, reply) => {
     const paramsSchema = z.object({
       id: z.string().uuid(),
     });
@@ -31,6 +38,10 @@ export async function memoriesRoutes(app: FastifyInstance) {
         memoryId: id,
       },
     });
+
+    if (!memory.isPublic && memory.userId !== request.user.sub) {
+      throw reply.status(401).send();
+    }
 
     return memory;
   });
@@ -48,13 +59,13 @@ export async function memoriesRoutes(app: FastifyInstance) {
         coverUrl,
         content,
         isPublic,
-        userId: "1",
+        userId: request.user.sub,
       },
     });
     return memory;
   });
 
-  app.put("/memories/:id", async (request) => {
+  app.put("/memories/:id", async (request, reply) => {
     const paramsSchema = z.object({
       id: z.string().uuid(),
     });
@@ -68,7 +79,17 @@ export async function memoriesRoutes(app: FastifyInstance) {
     });
     const { coverUrl, content, isPublic } = bodySchema.parse(request.body);
 
-    const memory = await prisma.memory.update({
+    let memory = await prisma.memory.findUniqueOrThrow({
+      where: {
+        memoryId: id,
+      },
+    });
+
+    if (memory?.userId !== request.user.sub) {
+      throw reply.status(401).send();
+    }
+
+    memory = await prisma.memory.update({
       where: {
         memoryId: id,
       },
@@ -82,12 +103,22 @@ export async function memoriesRoutes(app: FastifyInstance) {
     return memory;
   });
 
-  app.delete("/memories/:id", async (request) => {
+  app.delete("/memories/:id", async (request, reply) => {
     const paramsSchema = z.object({
       id: z.string().uuid(),
     });
 
     const { id } = paramsSchema.parse(request.params);
+
+    const memory = await prisma.memory.findUniqueOrThrow({
+      where: {
+        memoryId: id,
+      },
+    });
+
+    if (memory?.userId !== request.user.sub) {
+      throw reply.status(401).send();
+    }
 
     await prisma.memory.delete({
       where: {
